Little refactor
c-interface renamed to binding.
